backend/requirements.txt
django==2.2
django-bootstrap4
cryptography
djangorestframework
pillow
pyAesCrypt
